pyproject.toml
[project]
name = "cohort5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.8.3",
    "pydantic-ai-slim[openai]>=1.1.0",
    "python-dotenv>=1.1.1",
]
